requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
manim>=0.18.0
python-dotenv>=1.0.0
